+++ fvdb_reality_capture/sfm_scene/_load_e57_scene.py
@@ -8,7 +8,7 @@ import cv2
 import numpy as np
 import pye57
 import tqdm
-from fvdb import CameraModel
+from ..enums import CameraModel
 from scipy.spatial.transform import Rotation
 
 from . import SfmCache, SfmCameraMetadata, SfmPosedImageMetadata